removed 'section' files
