--- components/AboutSection.tsx
@@ -1,9 +1,0 @@
-import { Section } from '.'
-
-export const AboutSection = () => {
-  return (
-    <Section id='about' header='About Me'>
-      <p>Coming soon...</p>
-    </Section>
-  )
-}
