+++ components/index.ts
@@ -1,8 +1,7 @@
-export * from './AboutSection'
-export * from './GallerySection'
+export * from './DesktopGallery'
+export * from './MobileGallery'
 export * from './Input'
-export * from  './Layout'
+export * from './Layout'
 export * from './Datepicker'
 export * from './Navigation'
-export * from './OrderSection'
-export * from './Section'
+export * from './Page'
