+++ package.json
@@ -37,6 +37,7 @@
     "eslint": "8.24.0",
     "eslint-config-next": "12.3.1",
     "postcss": "^8.4.16",
+    "prettier": "^3.0.3",
     "tailwindcss": "^3.1.8",
     "typescript": "4.8.3"
   }
